requirements.txt
Django<1.9
Pillow==2.7.0
dj-database-url==0.3.0
-e git+https://github.com/pennersr/django-allauth@900b10957f7a805800a6153bd526f28666d28be1
django-courriers==0.5.18
django-extensions==1.5.5
django-formwizard==1.0
django-rosetta==0.7.6
django-separatedvaluesfield==0.2
django-toolbelt==0.0.1
django-widget-tweaks==1.3
docopt==0.4.0
gunicorn==19.1.1
healthgraph-api==0.3.0
mailchimp==2.0.9
psycopg2==2.6.1
python-dateutil==2.4.2
pytz==2014.10
requests==2.7.0
static3==0.5.1
stripe==1.23.0
waitress==0.8.9
whitenoise==1.0.6
wsgiref==0.1.2
hiredis==0.2.0
django-redis==4.2.0
redis==2.10.3
celery==3.1.18
django-bootstrap3==6.1.0
django_nose
coverage
